requirements.txt
fastapi
uvicorn
python-dotenv
cerebras-cloud-sdk
